refactor: add id params in getPhotos request
